backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-Cors
SQLAlchemy
psycopg2-binary
requests
python-dotenv
google-auth
